use auto-sync feature for question related API request
