+++ src/main/java/teammates/ui/webapi/action/CreateFeedbackQuestionAction.java
@@ -8,6 +8,7 @@ import teammates.common.exception.InvalidHttpRequestBodyException;
 import teammates.common.exception.InvalidParametersException;
 import teammates.common.util.Const;
 import teammates.ui.webapi.output.FeedbackQuestionData;
+import teammates.ui.webapi.request.FeedbackQuestionCreateRequest;
 
 /**
  * Create a feedback question.
@@ -35,8 +36,7 @@ public class CreateFeedbackQuestionAction extends Action {
         String courseId = getNonNullRequestParamValue(Const.ParamsNames.COURSE_ID);
         String feedbackSessionName = getNonNullRequestParamValue(Const.ParamsNames.FEEDBACK_SESSION_NAME);
 
-        FeedbackQuestionInfo.FeedbackQuestionCreateRequest request =
-                getAndValidateRequestBody(FeedbackQuestionInfo.FeedbackQuestionCreateRequest.class);
+        FeedbackQuestionCreateRequest request = getAndValidateRequestBody(FeedbackQuestionCreateRequest.class);
         FeedbackQuestionAttributes attributes = FeedbackQuestionAttributes.builder()
                 .withCourseId(courseId)
                 .withFeedbackSessionName(feedbackSessionName)
