+++ src/main/java/teammates/ui/webapi/request/FeedbackQuestionCreateRequest.java
@@ -1,0 +1,7 @@
+package teammates.ui.webapi.request;
+
+/**
+ * The create request of a feedback question.
+ */
+public class FeedbackQuestionCreateRequest extends FeedbackQuestionBasicRequest {
+}
