+++ src/main/java/teammates/ui/webapi/request/FeedbackQuestionSaveRequest.java
@@ -1,0 +1,7 @@
+package teammates.ui.webapi.request;
+
+/**
+ * The save request of a feedback question.
+ */
+public class FeedbackQuestionSaveRequest extends FeedbackQuestionBasicRequest {
+}
