+++ src/test/java/teammates/test/cases/webapi/SaveFeedbackQuestionActionTest.java
@@ -20,12 +20,12 @@ import teammates.common.exception.InvalidHttpRequestBodyException;
 import teammates.common.util.Const;
 import teammates.common.util.JsonUtils;
 import teammates.storage.api.FeedbackResponsesDb;
-import teammates.ui.webapi.action.FeedbackQuestionInfo;
 import teammates.ui.webapi.action.JsonResult;
 import teammates.ui.webapi.action.SaveFeedbackQuestionAction;
 import teammates.ui.webapi.output.FeedbackQuestionData;
 import teammates.ui.webapi.output.FeedbackVisibilityType;
 import teammates.ui.webapi.output.NumberOfEntitiesToGiveFeedbackToSetting;
+import teammates.ui.webapi.request.FeedbackQuestionSaveRequest;
 
 /**
  * SUT: {@link SaveFeedbackQuestionAction}.
@@ -62,7 +62,7 @@ public class SaveFeedbackQuestionActionTest extends BaseActionTest<SaveFeedbackQ
         String[] param = new String[] {
                 Const.ParamsNames.FEEDBACK_QUESTION_ID, typicalQuestion.getFeedbackQuestionId()
         };
-        FeedbackQuestionInfo.FeedbackQuestionSaveRequest saveRequest = getTypicalTextQuestionSaveRequest();
+        FeedbackQuestionSaveRequest saveRequest = getTypicalTextQuestionSaveRequest();
 
         SaveFeedbackQuestionAction a = getAction(saveRequest, param);
         JsonResult r = getJsonResult(a);
@@ -120,7 +120,7 @@ public class SaveFeedbackQuestionActionTest extends BaseActionTest<SaveFeedbackQ
         String[] param = new String[] {
                 Const.ParamsNames.FEEDBACK_QUESTION_ID, typicalQuestion.getFeedbackQuestionId()
         };
-        FeedbackQuestionInfo.FeedbackQuestionSaveRequest saveRequest = getTypicalTextQuestionSaveRequest();
+        FeedbackQuestionSaveRequest saveRequest = getTypicalTextQuestionSaveRequest();
         saveRequest.setNumberOfEntitiesToGiveFeedbackToSetting(NumberOfEntitiesToGiveFeedbackToSetting.CUSTOM);
         saveRequest.setCustomNumberOfEntitiesToGiveFeedbackTo(10);
 
@@ -145,7 +145,7 @@ public class SaveFeedbackQuestionActionTest extends BaseActionTest<SaveFeedbackQ
         String[] param = new String[] {
                 Const.ParamsNames.FEEDBACK_QUESTION_ID, typicalQuestion.getFeedbackQuestionId()
         };
-        FeedbackQuestionInfo.FeedbackQuestionSaveRequest saveRequest = getTypicalTextQuestionSaveRequest();
+        FeedbackQuestionSaveRequest saveRequest = getTypicalTextQuestionSaveRequest();
         saveRequest.setGiverType(FeedbackParticipantType.STUDENTS);
         saveRequest.setRecipientType(FeedbackParticipantType.TEAMS);
         saveRequest.setShowResponsesTo(Arrays.asList(FeedbackVisibilityType.RECIPIENT));
@@ -177,7 +177,7 @@ public class SaveFeedbackQuestionActionTest extends BaseActionTest<SaveFeedbackQ
         String[] param = new String[] {
                 Const.ParamsNames.FEEDBACK_QUESTION_ID, typicalQuestion.getFeedbackQuestionId()
         };
-        FeedbackQuestionInfo.FeedbackQuestionSaveRequest saveRequest = getTypicalTextQuestionSaveRequest();
+        FeedbackQuestionSaveRequest saveRequest = getTypicalTextQuestionSaveRequest();
         saveRequest.setGiverType(FeedbackParticipantType.STUDENTS);
         saveRequest.setRecipientType(FeedbackParticipantType.SELF);
         saveRequest.setShowResponsesTo(Arrays.asList(FeedbackVisibilityType.RECIPIENT));
@@ -216,7 +216,7 @@ public class SaveFeedbackQuestionActionTest extends BaseActionTest<SaveFeedbackQ
         // There are already responses for this question
         assertFalse(frDb.getFeedbackResponsesForQuestion(fq.getId()).isEmpty());
 
-        FeedbackQuestionInfo.FeedbackQuestionSaveRequest saveRequest = getTypicalContributionQuestionSaveRequest();
+        FeedbackQuestionSaveRequest saveRequest = getTypicalContributionQuestionSaveRequest();
         saveRequest.setQuestionNumber(fq.getQuestionNumber());
         saveRequest.setGiverType(fq.getGiverType());
         saveRequest.setRecipientType(fq.getRecipientType());
@@ -236,7 +236,7 @@ public class SaveFeedbackQuestionActionTest extends BaseActionTest<SaveFeedbackQ
         ______TS("Edit: Invalid recipient type");
 
         assertThrows(InvalidHttpRequestBodyException.class, () -> {
-            FeedbackQuestionInfo.FeedbackQuestionSaveRequest request = getTypicalContributionQuestionSaveRequest();
+            FeedbackQuestionSaveRequest request = getTypicalContributionQuestionSaveRequest();
             request.setQuestionNumber(fq.getQuestionNumber());
             request.setRecipientType(FeedbackParticipantType.STUDENTS);
             getJsonResult(getAction(request, param));
@@ -255,7 +255,7 @@ public class SaveFeedbackQuestionActionTest extends BaseActionTest<SaveFeedbackQ
         String[] param = new String[] {
                 Const.ParamsNames.FEEDBACK_QUESTION_ID, typicalQuestion.getFeedbackQuestionId()
         };
-        FeedbackQuestionInfo.FeedbackQuestionSaveRequest saveRequest = getTypicalTextQuestionSaveRequest();
+        FeedbackQuestionSaveRequest saveRequest = getTypicalTextQuestionSaveRequest();
         saveRequest.setQuestionNumber(-1);
 
         SaveFeedbackQuestionAction a = getAction(saveRequest, param);
@@ -281,7 +281,7 @@ public class SaveFeedbackQuestionActionTest extends BaseActionTest<SaveFeedbackQ
         String[] param = new String[] {
                 Const.ParamsNames.FEEDBACK_QUESTION_ID, typicalQuestion.getFeedbackQuestionId()
         };
-        FeedbackQuestionInfo.FeedbackQuestionSaveRequest saveRequest = getTypicalTextQuestionSaveRequest();
+        FeedbackQuestionSaveRequest saveRequest = getTypicalTextQuestionSaveRequest();
         saveRequest.setGiverType(FeedbackParticipantType.TEAMS);
         saveRequest.setRecipientType(FeedbackParticipantType.OWN_TEAM_MEMBERS);
 
@@ -322,7 +322,7 @@ public class SaveFeedbackQuestionActionTest extends BaseActionTest<SaveFeedbackQ
 
         FeedbackQuestionAttributes fq =
                 logic.getFeedbackQuestion(fs.getFeedbackSessionName(), fs.getCourseId(), 1);
-        FeedbackQuestionInfo.FeedbackQuestionSaveRequest saveRequest = getTypicalTextQuestionSaveRequest();
+        FeedbackQuestionSaveRequest saveRequest = getTypicalTextQuestionSaveRequest();
         saveRequest.setQuestionNumber(fq.getQuestionNumber());
         saveRequest.setGiverType(FeedbackParticipantType.STUDENTS);
         saveRequest.setRecipientType(FeedbackParticipantType.STUDENTS);
@@ -387,9 +387,8 @@ public class SaveFeedbackQuestionActionTest extends BaseActionTest<SaveFeedbackQ
         assertEquals(totalStudents + 1, details.stats.expectedTotal);
     }
 
-    private FeedbackQuestionInfo.FeedbackQuestionSaveRequest getTypicalTextQuestionSaveRequest() {
-        FeedbackQuestionInfo.FeedbackQuestionSaveRequest saveRequest =
-                new FeedbackQuestionInfo.FeedbackQuestionSaveRequest();
+    private FeedbackQuestionSaveRequest getTypicalTextQuestionSaveRequest() {
+        FeedbackQuestionSaveRequest saveRequest = new FeedbackQuestionSaveRequest();
         saveRequest.setQuestionNumber(2);
         saveRequest.setQuestionBrief("this is the brief");
         saveRequest.setQuestionDescription("this is the description");
@@ -408,9 +407,8 @@ public class SaveFeedbackQuestionActionTest extends BaseActionTest<SaveFeedbackQ
         return saveRequest;
     }
 
-    private FeedbackQuestionInfo.FeedbackQuestionSaveRequest getTypicalContributionQuestionSaveRequest() {
-        FeedbackQuestionInfo.FeedbackQuestionSaveRequest saveRequest =
-                new FeedbackQuestionInfo.FeedbackQuestionSaveRequest();
+    private FeedbackQuestionSaveRequest getTypicalContributionQuestionSaveRequest() {
+        FeedbackQuestionSaveRequest saveRequest = new FeedbackQuestionSaveRequest();
         saveRequest.setQuestionNumber(1);
         saveRequest.setQuestionBrief("this is the brief for contribution question");
         saveRequest.setQuestionDescription("this is the description for contribution question");
