+++ src/web/app/components/question-types/question-types-session-submission/question-submission-form.component.spec.ts
@@ -1,5 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
+import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { QuestionSubmissionFormComponent } from './question-submission-form.component';
 import { QuestionTypesSessionSubmissionModule } from './question-types-session-submission.module';
 
@@ -11,6 +12,7 @@ describe('QuestionSubmissionFormComponent', () => {
     TestBed.configureTestingModule({
       imports: [
         QuestionTypesSessionSubmissionModule,
+        HttpClientTestingModule,
       ],
     })
     .compileComponents();
